example/requirements.txt
django
celery
beautifulsoup
django-celery
python-memcached
